Scan is working quite well

+++ Walnut-AT/cli_commands.h
@@ -1,6 +1,6 @@
 #ifndef CLI_COMMANDS_
 #define CLI_COMMANDS_
 
-void execute_cli_command(char *str, int strlen);
+void execute_cli_command(char *str, uint8_t strlen);
 
 #endif  // CLI_COMMANDS_
